// frontend/generated/models/ReportDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AccountDto } from './AccountDto';
import {
    AccountDtoFromJSON,
    AccountDtoFromJSONTyped,
    AccountDtoToJSON,
    AccountDtoToJSONTyped,
} from './AccountDto';

/**
 * 
 * @export
 * @interface ReportDto
 */
export interface ReportDto {
    /**
     * 
     * @type {number}
     * @memberof ReportDto
     */
    reportId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof ReportDto
     */
    reporter?: AccountDto;
    /**
     * 
     * @type {number}
     * @memberof ReportDto
     */
    projectId?: number;
    /**
     * 
     * @type {string}
     * @memberof ReportDto
     */
    reason?: string;
    /**
     * 
     * @type {string}
     * @memberof ReportDto
     */
    solution?: string;
    /**
     * 
     * @type {string}
     * @memberof ReportDto
     */
    status?: ReportDtoStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof ReportDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ReportDto
     */
    updatedAt?: Date;
}


/**
 * @export
 */
export const ReportDtoStatusEnum = {
    Unsolved: 'UNSOLVED',
    Solved: 'SOLVED',
    Solving: 'SOLVING'
} as const;
export type ReportDtoStatusEnum = typeof ReportDtoStatusEnum[keyof typeof ReportDtoStatusEnum];


/**
 * Check if a given object implements the ReportDto interface.
 */
export function instanceOfReportDto(value: object): value is ReportDto {
    return true;
}

export function ReportDtoFromJSON(json: any): ReportDto {
    return ReportDtoFromJSONTyped(json, false);
}

export function ReportDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReportDto {
    if (json == null) {
        return json;
    }
    return {
        
        'reportId': json['reportId'] == null ? undefined : json['reportId'],
        'reporter': json['reporter'] == null ? undefined : AccountDtoFromJSON(json['reporter']),
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
        'reason': json['reason'] == null ? undefined : json['reason'],
        'solution': json['solution'] == null ? undefined : json['solution'],
        'status': json['status'] == null ? undefined : json['status'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function ReportDtoToJSON(json: any): ReportDto {
    return ReportDtoToJSONTyped(json, false);
}

export function ReportDtoToJSONTyped(value?: ReportDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'reportId': value['reportId'],
        'reporter': AccountDtoToJSON(value['reporter']),
        'projectId': value['projectId'],
        'reason': value['reason'],
        'solution': value['solution'],
        'status': value['status'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}